+++ src/Rest/Routes/Integrations/Calculator/FormSubmitCalculatorRoute.php
@@ -61,6 +61,7 @@ class FormSubmitCalculatorRoute extends AbstractIntegrationFormSubmit
 		return [
 			Config::FD_FORM_ID => 'string',
 			Config::FD_POST_ID => 'string',
+			Config::FD_PARAMS => 'array',
 		];
 	}