+++ src/Rest/Routes/Integrations/Hubspot/FormSubmitHubspotRoute.php
@@ -116,6 +116,7 @@ class FormSubmitHubspotRoute extends AbstractIntegrationFormSubmit
 			Config::FD_FORM_ID => 'string',
 			Config::FD_POST_ID => 'string',
 			Config::FD_ITEM_ID => 'string',
+			Config::FD_PARAMS => 'array',
 		];
 	}