+++ src/Rest/Routes/Integrations/Mailer/FormSubmitMailerRoute.php
@@ -57,6 +57,7 @@ class FormSubmitMailerRoute extends AbstractIntegrationFormSubmit
 		return [
 			Config::FD_FORM_ID => 'string',
 			Config::FD_POST_ID => 'string',
+			Config::FD_PARAMS => 'array',
 		];
 	}