+++ src/Rest/Routes/Integrations/Moments/FormSubmitMomentsRoute.php
@@ -117,6 +117,7 @@ class FormSubmitMomentsRoute extends AbstractIntegrationFormSubmit
 			Config::FD_FORM_ID => 'string',
 			Config::FD_POST_ID => 'string',
 			Config::FD_ITEM_ID => 'string',
+			Config::FD_PARAMS => 'array',
 		];
 	}