+++ src/Rest/Routes/Integrations/Pipedrive/FormSubmitPipedriveRoute.php
@@ -104,6 +104,7 @@ class FormSubmitPipedriveRoute extends AbstractIntegrationFormSubmit
 		return [
 			Config::FD_FORM_ID => 'string',
 			Config::FD_POST_ID => 'string',
+			Config::FD_PARAMS => 'array',
 		];
 	}